+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useModal } from '../Modal/useModal';
-import Modal from '../Modal/Modal';
+import CreateRestaurantModal from '../Modal/CreateRestaurantModal';
 
 function Header() {
   const { openModal, closeModal, isOpen } = useModal();
@@ -15,35 +15,7 @@ function Header() {
         </button>
       </header>
       {/* Modal */}
-      <Modal isOpen={isOpen} closeModal={closeModal}>
-        <Modal.Title>새로운 음식점</Modal.Title>
-        <div className='flex flex-col gap-2'>
-          <label htmlFor='name'>음식점 이름</label>
-          <input className='w-full rounded-md border-2 border-gray-300 p-2' type='text' id='name' />
-          <label htmlFor='address'>주소</label>
-          <input
-            className='w-full rounded-md border-2 border-gray-300 p-2'
-            type='text'
-            id='address'
-          />
-          <label htmlFor='phone'>전화번호</label>
-          <input
-            className='w-full rounded-md border-2 border-gray-300 p-2'
-            type='text'
-            id='phone'
-          />
-          <label htmlFor='category'>카테고리</label>
-          <input type='text' id='category' />
-        </div>
-        <Modal.Footer>
-          <Modal.Button colorType='white' onClick={closeModal}>
-            취소
-          </Modal.Button>
-          <Modal.Button colorType='orange' onClick={closeModal}>
-            추가
-          </Modal.Button>
-        </Modal.Footer>
-      </Modal>
+      <CreateRestaurantModal isOpen={isOpen} closeModal={closeModal} />
     </>
   );
 }
