+++ src/constants/resaurant.ts
@@ -1,0 +1,10 @@
+const CATEGORY_IMAGE_MAP = {
+  한식: './category-korean.png',
+  중식: './category-chinese.png',
+  일식: './category-japanese.png',
+  양식: './category-western.png',
+  아시안: './category-asian.png',
+  기타: './category-etc.png',
+};
+
+export { CATEGORY_IMAGE_MAP };
